Removed kotlin reflection from main module

+++ src/main/kotlin/org/tesserakt/diskordin/core/data/json/request/GuildRoleCreateRequest.kt
@@ -3,7 +3,7 @@ package org.tesserakt.diskordin.core.data.json.request
 
 data class GuildRoleCreateRequest(
     val name: String? = null,
-    val permissions: Int? = null,
+    val permissions: Long? = null,
     val color: Int? = null,
     val hoist: Boolean? = null,
     val mentionable: Boolean? = null
